fix: don't query workspace in `UpdateWorkspaceAgentConnectionByID`
